Add support for ?tls=0 query parameter
This parameter can be used to opt out of TLS when using `libsql:` URL.
It is also possible to use it with HTTP and WebSocket URLs, but in this
case we just check that it matches the existing scheme (i.e.,
`https://localhost?tls=0` throws an error). We also require every
`libsql:` URL with `?tls=0` to specify an explicit port, because the
default port for `libsql:` URL is 443, which could produce unexpected
results.

This also makes sure that we don't send the parsed query parameters
(such as the auth token) to the server. The code apparently attempted to
do so by calling `url.Query().Del()`, but this modifies only a temporary
copy of the parsed query parameters, not the original URL.