bugfix: propagate replication index to guarantee read your writes (#103)
Add `ReplicationIndex` field to relevant structs

Added the `ReplicationIndex` to `Batch`, `BatchResult`, `Stmt` and `StmtResult`. The field type is *uint64.

The libsql-server expects this field as a string type. So, a custom json marshaler is written for it.

The libsql-server earlier used to send this field as uint64 which is now changed to string. To accomodate both, a custom json unmarshaler is written to handle both the cases.