Optimize arguments handling
Remove memory/cpu overhead of handling arguments when they are empty.
This is important in case of user loading a dump of their database.
In such a case, there are no arguments at all and the number of
statements is usually big.

When testing on a 650M sql file this change reduces number of total
allocation from 122772583kB to 38511376 so ~69%.

Signed-off-by: Piotr Jastrzebski <piotr@chiselstrike.com>